requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
Django==1.11.4
django-bootstrap-datepicker==1.2.3
django-forms-bootstrap==3.1.0
idna==2.6
pytz==2017.2
requests==2.18.4
urllib3==1.22
